+++ joblib/test/test_numpy_pickle.py
@@ -279,13 +279,15 @@ def test_masked_array_persistence(tmpdir):
 def test_compress_mmap_mode_warning(tmpdir):
     # Test the warning in case of compress + mmap_mode
     rnd = np.random.RandomState(0)
-    a = rnd.random_sample(10)
+    obj = rnd.random_sample(10)
     this_filename = tmpdir.join("test.pkl").strpath
-    numpy_pickle.dump(a, this_filename, compress=1)
+    numpy_pickle.dump(obj, this_filename, compress=1)
     with warns(UserWarning) as warninfo:
-        numpy_pickle.load(this_filename, mmap_mode="r+")
+        reloaded_obj = numpy_pickle.load(this_filename, mmap_mode="r+")
     debug_msg = "\n".join([str(w) for w in warninfo])
     warninfo = [w.message for w in warninfo]
+    assert not isinstance(reloaded_obj, np.memmap)
+    np.testing.assert_array_equal(obj, reloaded_obj)
     assert len(warninfo) == 1, debug_msg
     assert (
         str(warninfo[0]) == 'mmap_mode "r+" is not compatible with compressed '